refactor: split logic to run and parse commands

+++ src/packages.h
@@ -8,6 +8,7 @@
 #pragma once
 
 #include <stddef.h>
+#include <stdint.h>
 
 #define LPM_PACKAGE_STATUS_INSTALLED "[*]"
 #define LPM_PACKAGE_STATUS_AVAILABLE "[-]"
@@ -27,4 +28,4 @@ typedef struct
 } LPM_Packages;
 
 void lpm_packages_teardown(LPM_Packages *pkgs);
-int lpm_packages_get(LPM_Packages *pkgs, const char *pkg_name);
+uint8_t lpm_packages_get(LPM_Packages *pkgs, const char *pkg_name);
